backend/utils/ai_models.py
import torch
from torchvision import transforms
from PIL import Image
import io
import os
import requests

class AIModels:
    def __init__(self):
        self.device = torch.device("cuda" if torch.cuda.is_available() else "cpu")
        self.colorizer_model = None
        self.enhancer_model = None
        
    def load_colorizer(self, model_path="models/colorizer.pth"):
        # Placeholder for loading DeOldify or similar
        # In a real scenario, we would load the weights here
        if os.path.exists(model_path):
            # self.colorizer_model = torch.load(model_path, map_location=self.device)
            pass
        return "Colorizer loaded (Placeholder)"

    def load_enhancer(self, model_path="models/enhancer.pth"):
        # Placeholder for Real-ESRGAN
        if os.path.exists(model_path):
            pass
        return "Enhancer loaded (Placeholder)"

    async def colorize(self, image_bytes):
        # Implementation logic for colorization
        # 1. Byte to Image
        # 2. Preprocess (Resize to 512, Normalize)
        # 3. Model Inference
        # 4. Postprocess (Upscale back to original, Blend)
        return image_bytes # Currently returning original for placeholder

    async def sharpen(self, image_bytes):
        # Implementation logic for Real-ESRGAN
        return image_bytes
