backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
opencv-python-headless
Pillow
torch
torchvision
celery[redis]
redis
pydantic
pydantic-settings
python-dotenv
